gui: Fix disk space calculation on some platforms
The setup wizard needs to know how much disk is available
on the user's system so it can inform the user if there
is insufficient disk space to download the (rather large)
Granite models.

The code currently does this using the systeminformation
module to check the available space on the root filesystem.

That isn't exactly correct, though, because models are
stored in the users directory, which may be on a different
filesystem altogether.

Furthermore, systeminformation doesn't handle btrfs systems
properly, since it keys output off the volume name, and the
volume name is shared between subvolumes with btrfs.

This commit addresses both problems by switching the code
over to use the check-disk-space module, which is more
form fit for our purposes anyway.

Signed-off-by: Fred Bricon <fbricon@gmail.com>