gui: Deduct existing models from diskspace needs
The setup wizard reports to the user if there is insufficent
diskspace to download the Granite model the user has selected.

If there is insufficient diskspace to proceed it stops the
user from continuing.

Unfortunately, there is a oversight in the code where it fails
to consider that the user may already have a model downloaded.

In that case, the model doesn't need new diskspace.

This commit fixes that problem by reporting the download status
of each model to the wizard and then ignoring already downloaded
models when calculating diskspace needs.

Signed-off-by: Fred Bricon <fbricon@gmail.com>